+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 unreleased 
  - allow dmatrix for prediction
-
+ - made `scicloj.metamorph.ml.viz/apply-xform-kvs` public
+ 
 1.5.1
  - added `r-model-matrix/lm` helper for one-line linear models
  - added ns documentation to most ns